### 2019-07-07 #1551 @whedon
[PRE REVIEW]: studentlife: Tidy Handling and Navigation of a Valuable Mobile-Health Dataset
**Submitting author:** @frycast (<a href="http://orcid.org/0000-0001-6032-0522">Daniel Fryer</a>)
**Repository:** <a href="github.com/frycast/studentlife" target ="_blank">https://github.com/frycast/studentlife</a>
**Version:** v1.0.0
