### 2019-07-07 #1551 @labarba
@arokem — the author requested you as handling editor, but I see that your load is on the high side right now. Are you willing to take another one? I can look for someone else if all your papers are mid-flight. Let me know!
