### 2019-07-07 #1551 @labarba
I think this is the correct URL — pls. confirm @ frycast
https://github.com/frycast/studentlife
