### 2019-07-07 #1551 @arokem
@frycast, as mentioned by whedon [above ](https://github.com/openjournals/joss-reviews/issues/1551#issuecomment-509014714), the paper, currently under a `JOSS-paper` folder can't be automatically found in this location. Could you please move it to a folder called `paper`?
