### 2019-07-07 #1551 @arokem
:wave: @tgerke : would you be willing to review this paper for JOSS? Your expertise in public health and R would be helpful here.

:wave: @pacoramon : would be willing to review this paper for JOSS? Your expertise in health and medicine, as well as statistical software would be helpful here.

:wave: @borishejblum : would you be willing to review this paper for JOSS? Your expertise in epidemiology and R would be helpful here.
