### 2019-07-07 #1551 @labarba
I thought it couldn't be found because the GitHub URL provided was wrong, and once you fixed that URL `whedon` would find it?
