### 2019-07-07 #1551 @frycast
I tried to generate pdf with paper files in folder 'paper' and then I tried to generate pdf with them in the package root. I've now moved them back into folder 'paper'.
