### 2019-07-12 #1551 @arokem
:wave: @corinne-riddell : would be willing to review this paper for JOSS? Your expertise in epidemiology and R would be helpful here.

:wave: @jenniferthompson : would be willing to review this paper for JOSS? Your expertise in public health and R would be helpful here.

:wave: @Jminnier : would be willing to review this paper for JOSS? Your expertise in biostatistics and R would be helpful here
