### 2019-07-17 #1551 @arokem
👋 @Rmadillo : would be willing to review this paper for JOSS? Your expertise in health informatics and R would be helpful here.

👋 @klmedeiros : would you be willing to review this paper for JOSS? Your expertise in public health, epi and R would be helpful here.

👋 @acolum : would be willing to review this paper for JOSS? Your expertise in public health, epidemiology and R would be helpful here.
