### 2019-07-21 #1587 @whedon
[REVIEW]: studentlife: Tidy Handling and Navigation of a Valuable Mobile-Health Dataset
**Submitting author:** @frycast (<a href="http://orcid.org/0000-0001-6032-0522">Daniel Fryer</a>)
**Repository:** <a href="https://github.com/frycast/studentlife" target ="_blank">https://github.com/frycast/studentlife</a>
**Version:** v1.0.0

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the <a href="github.com/frycast/studentlife" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

@klmedeiros, @jminnier : please carry out your review in this issue by updating your checklist below. If you cannot edit the checklist please:

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="github.com/frycast/studentlife">repository url</a>?

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="github.com/frycast/studentlife">repository url</a>?
