### 2019-07-21 #1551 @arokem
@jminnier, @klmedeiros : the review will take place in #1587. Please head over there for a checklist of the items that we would like you to review and to post your comments. See also https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for guidelines.
