### 2019-07-22 #1587 @hiendn
> **Submitting author:** @frycast ([Daniel Fryer](http://orcid.org/0000-0001-6032-0522))
> **Repository:** [github.com/frycast/studentlife](github.com/frycast/studentlife)
> **Version:** v1.0.0
> **Editor:** @arokem
> **Reviewers:** @klmedeiros, @jminnier
> **Archive:** Pending
>
> ## Status
> [![status](https://camo.githubusercontent.com/cee88cbf149132c3f25c18f7c74e1e4db7b29829/687474703a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f34613631643034346165666162646237633537353230613037313565343165642f7374617475732e737667)](http://joss.theoj.org/papers/4a61d044aefabdb7c57520a0715e41ed)
>
> Status badge code:
>
> ```
> ```
>
> **Reviewers and authors**:
>
> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](github.com/frycast/studentlife) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
>
> ## Reviewer instructions & questions
> @klmedeiros, @jminnier : please carry out your review in this issue by updating your checklist below. If you cannot edit the checklist please:
>
> 1. Make sure you're logged in to your GitHub account
> 2. Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations
>
> The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let @arokem know.
>
> ✨ **Please try and complete your review in the next two weeks** ✨
>
> ## Review checklist for @klmedeiros
> ### Conflict of interest
> * [ ]  As the reviewer I confirm that I have read the [JOSS conflict of interest policy](https://github.com/openjournals/joss/blob/master/COI.md) and that there are no conflicts of interest for me to review this work.
>
> ### Code of Conduct
> * [ ]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [ ]  **Repository:** Is the source code for this software available at the [repository url](github.com/frycast/studentlife)?
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [ ]  **Version:** Does the release version given match the GitHub release (v1.0.0)?
> * [ ]  **Authorship:** Has the submitting author (@frycast) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> ### Functionality
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
> * [ ]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?
> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> ### Software paper
> * [ ]  **Authors:** Does the `paper.md` file include a list of authors with their affiliations?
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
>
> ## Review checklist for @jminnier
> ### Conflict of interest
> * [ ]  As the reviewer I confirm that I have read the [JOSS conflict of interest policy](https://github.com/openjournals/joss/blob/master/COI.md) and that there are no conflicts of interest for me to review this work.
>
> ### Code of Conduct
> * [ ]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [ ]  **Repository:** Is the source code for this software available at the [repository url](github.com/frycast/studentlife)?
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [ ]  **Version:** Does the release version given match the GitHub release (v1.0.0)?
> * [ ]  **Authorship:** Has the submitting author (@frycast) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> ### Functionality
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
> * [ ]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?
> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> ### Software paper
> * [ ]  **Authors:** Does the `paper.md` file include a list of authors with their affiliations?
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

@arokem @frycast --- The URL at the top of the page leads to a 404 error when clicked. Is this a potential reason why the paper is not compiling by `@weldon generate pdf`?
