### 2019-07-31 #1587 @hiendn
@arokem can you look into what is wrong with this submission, that makes it unable to render a PDF? Do you think it is worth it to simply resubmit the paper in a hope of resolving the issues?
