### 2019-08-05 e4818ca
Update paper.md
Fix typo "monstreal"
