### 2019-08-05 #1587 @frycast
Hi @arokem  @jminner @klmedeiros. I've updated the typo in the author's affiliation and recompiled the pdf. How is it looking now?
