### 2019-08-06 frycast/studentlife#7 @jminnier
Typos
A couple minor things found when reviewing for JOSS https://github.com/openjournals/joss-reviews/issues/1587

- In the readme, I believe `download_studentlife(dest = d)` should be `download_studentlife(location = d)`
- In documentation of `add_block_labels()` argument `interval` "is decided when `stude` is of class `interval_SL_tibble`" -- what is `stude`? I think it's a typo...
