### 2019-08-06 frycast/studentlife#8 @jminnier
Add contribution info
One thing JOSS looks for is information on contribution/support, it would be useful to add a contribution instructions document that outlines this, or add this info to the readme:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

https://github.com/openjournals/joss-reviews/issues/1587
