### 2019-08-08 #1587 @jminnier
This software looks great and the addition of visualizations makes this data accessing package really nice and extra useful.

I just had minor questions:
https://github.com/frycast/studentlife/issues/7 (a couple typos, I think)
https://github.com/frycast/studentlife/issues/8 (contribution info missing as far as I can tell)

Thanks for making this package @frycast!
