### 2019-08-08 #1587 @arokem
Thanks for taking a look @jminnier !

@klmedeiros : have you had a chance to look into this?
