### 2019-08-09 edac3c2
fix typo in documentation "stude" -> "tab"
