### 2019-08-09 #1587 @frycast
Thanks for the useful comments @jminnier . I've fixed both typos and appended community guidelines to the README . Thanks @arokem .
