### 2019-08-16 #1587 @frycast
@klmedeiros : sorry for the delay in getting the pdf to compile in the beginning. Do you still have time to review this?
