### 2019-08-16 #1587 @klmedeiros
@frycast Yes my sincerest apologies for the delay on my end. I will wrap this up today/tomorrow.
