### 2019-08-19 #1587 @klmedeiros
@frycast This looks great! I found the documentation on the README very helpful, and was happy to provide a review for this.
