### 2019-08-19 #1587 @arokem
@klmedeiros : there is still one item that is not checked off ("performance"). Are there any performance claims that you would like to test? Otherwise, is this paper ready to be accepted from your point of view?
