### 2019-08-19 #1587 @arokem
@jminnier : just to confirm: is this paper ready to be accepted from your point of view?
