### 2019-08-19 #1587 @jminnier
@arokem Indeed, I believe this paper is ready to be accepted.
