### 2019-08-19 #1587 @klmedeiros
@arokem My apologies, I did test the package out a bit (differences between downloading full dataset or a sample) and found it good, and forgot to check that off. I have now, and I believe the paper is ready to be accepted.
