### 2019-08-20 34800a3
delete CRAN-RELEASE and update NEWS.md
