+++ NEWS.md
@@ -1,2 +1,9 @@
+# studentlife 1.0.1
+This release includes changes recommended by 
+reviewers at the Journal of Open Source Software.
+
+  * Fixed typos in documentation.
+  * Include community guidelines in the README.md
+
 # studentlife 1.0.0
 This is the first release.
