### 2019-08-20 3f0e249
Update version number in DESCRIPTION
