### 2019-08-20 e9579dd
updated authorship information in DESCRIPTION
