### 2019-08-20 655870e
create .zenodo.json and update version number again
