### 2019-08-20 #1587 @frycast
Thanks @arokem . I've made the new release and an archive on Zenodo with DOI: 10.5281/zenodo.3371922
