### 2019-08-20 #1587 @arokem
This looks good to me

@openjournals/joss-eics : I believe that this paper is ready for your review.
