### 2019-08-20 ec17a1f
Update paper.md
Minor copy editing of JOSS paper
