+++ paper/paper.md
@@ -32,9 +32,9 @@ bibliography: paper.bib
 ---
 
 # Summary
-Smartphones have become ubiquitous; and nearly constantly accompany their owners, as they go about their daily lives. This sensor-rich and computationally-efficient technology allows quantifying continuous streams of human behaviors and activities unobtrusively and *in situ*, a process referred to as digital phenotyping [@Onnela:2016]. Two main types of data collected through this approach can be distinguished: on the one hand, active data that involve the voluntary participation of users, as required for surveys; and on the other hand, passive data that do not involve any dedicated action from users, as is the case for data generated by various sensors embedded in the smartphones.
+Smartphones have become ubiquitous, and nearly constantly accompany their owners as they go about their daily lives. This sensor-rich and computationally-efficient technology allows quantifying continuous streams of human behaviors and activities unobtrusively and *in situ*, a process referred to as digital phenotyping [@Onnela:2016]. Two main types of data collected through this approach can be distinguished: on the one hand, active data that involve the voluntary participation of users, as required for surveys; and on the other hand, passive data that do not involve any dedicated action from users, as is the case for data generated by various sensors embedded in the smartphones.
 
-The R package [@RCore:2019] **``studentlife``** is designed to assist in downloading, navigating and analysing the **StudentLife** dataset. The **StudentLife** dataset [@Wang:2014] is a one-of-its-kind digital phenotyping dataset made publicly available (https://studentlife.cs.dartmouth.edu) by Wang, Campbell and colleagues, at Dartmouth College (New Hampshire, United States). It consists of a rich array of data, collected through a dedicated mobile app, over a 10-week term in a sample of 48 college students. Active data were obtained through ecological momentary assessments, corresponding to self-reports of the students’ states, prompted multiple times per day (3 to 13 daily assessments) at pseudorandom intervals. Likert scale-based questions (e.g., regarding stress, exercise, social activity, etc) that changed across assessments, were systematically preceded by a photographic affect meter that captured the students’ instantaneous mood in an original fashion, hence promoting the collection of a large amount of affect data. Passive data were also continuously collected through smartphone sensors (e.g., accelerometer, microphone, GPS, etc). A key valuable aspect of the **StudentLife** dataset is that it does not merely share raw sensor data. Indeed, a number of behavioral inferences based on trained machine learners were automatically computed by the mobile app and are also included in the dataset, which includes, for example, type of physical activity, audio/conversation state, and sleep duration. Finally, the dataset contains data from pre- and post-term surveys aimed at assessing mental health with a series of well-validated questionnaires and scales (e.g., PHQ-9 to screen for the presence and severity of depression), as well as a number of measures related to academic context and performance. 
[…]
