+++ paper/paper.bib
@@ -13,10 +13,11 @@
 @inproceedings{Wang:2014,
   title={{StudentLife: Assessing Mental Health, Academic Performance and Behavioral Trends of College Students Using Smartphones}},
   author={Wang, Rui and Chen, Fanglin and Chen, Zhenyu and Li, Tianxing and Harari, Gabriella and Tignor, Stefanie and Zhou, Xia and Ben-Zeev, Dror and Campbell, Andrew T},
-  booktitle={Proceedings of the ACM Conference on Ubiquitous Computing},
+  booktitle={Proceedings of the 2014 {ACM} {International} {Joint} {Conference} on {Pervasive} and {Ubiquitous} {Computing}},
   year={2014},
   organization={ACM},
-  doi={10.1145/2632048.2632054}
+  doi={10.1145/2632048.2632054},
+  pages={3--14},
 }
 
 @article{RCore:2019,
