### 2019-08-20 #1587 @kyleniemeyer
Hi @frycast, I made some minor copy edits in two PRs, if you could merge those: https://github.com/frycast/studentlife/pull/9 and https://github.com/frycast/studentlife/pull/10
