### 2019-08-20 #1587 @frycast
Thanks @kyleniemeyer . I've merged them both.
