### 2019-08-21 #1587 @kyleniemeyer
Congrats @frycast on your article's publication in JOSS! Many thanks to @arokem for editing, and @klmedeiros and @jminnier for reviewing this submission.
