### 2019-08-22 #1587 @frycast
@kyleniemeyer the software [repository link here](https://joss.theoj.org/papers/4a61d044aefabdb7c57520a0715e41ed) still leads to 404 Not Found page. It's because I originally specified the link as "github.com/frycast/studentlife" and it needs to be changed to have the prefix "https://". Are you able to make that change?
